Masters Projects
This project is a reliable web/mobile application which provides a platform where the users ( people who want to send a courier from one location to another ) be able to send couriers by posting offers to drivers by providing price, information about the courier, pickup and drop-off date and time and pickup and drop-off location on google maps along with other details. The users have the option of becoming a driver. They would just require providing driver license details along with a clear copy of driver license in-order for the admin to verify them.
The drivers can accept the offers made by the users and there are no constraints on how many offers can a driver accept. The offers are accepted in FCFS method. Once an offer is accepted by a driver, the pickup and the delivery time gets stored in the database, no other offer will be displayed to the driver due to the Time Slot Algorithm used which automatically filter and removes out all those offers whose pickup or delivery time overlaps the already accepted offers pickup and delivery time. There are no constraints on the number of offers a driver can accept.
The drivers are also normal users of the system as well. The drivers also have the option of posting offers. The offers posted by a driver is not displayed to themselves as the Time Slot Algorithm filters those offers out.